+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## [unreleased]
+### Added
+- Functionality to get objects in a given time range
+
 ## 1.0.0
 
 - Initial version, created by Julian Sanchez